requirements.txt
polars>=1.21.0
requests>=2.31.0
pandas>=2.2.0
beautifulsoup4>=4.12.3
tqdm>=4.66.1
xlrd>=2.0.1
nltk>=3.8.1
google-generativeai>=0.3.0
python-dotenv>=1.0.0
openai>=1.43.0 